refactor: slightly less coupling in PreHandleMSG
We need to check for dark mode while the native window's hwnd is being
created. The WM_NCCREATE message is a reasonable place to do that, but
NativeWindowViews doesn't have access to the hwnd when that message is
sent.

Since DesktopWindowTreeHostWin _does_ have access to the hwnd, handle
the dark mode bootstrap check there instead of NativeWindowViews.